+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 > Committed but unreleased changes are put here, at the top. Older releases are detailed chronologically below. 
 
+
+
+## 2.21.12 (2024-07-20)
+
+#### Fixed
+
+- `dropdown` - Always drops down when not clipping. Before, this component would choose a body position closest to the vertical center of the viewport. Now, it chooses a position below the anchor, unless that would cause the body to clip.
+- `simple-v-table` - Renamed `:row-export-fn` to `:on-export-row-label-fn`.
+
+
 ## 2.21.11 (2024-07-11)
 
 #### Fixed